+++ src/server/types.ts
@@ -17,8 +17,6 @@ export type GraphNode = {
 export type Graph = Map<string, GraphNode>;
 
 declare global {
-  /* eslint-disable no-var */
   var __rds_start: number;
   var __rds_profile_session: Session | undefined;
-  /* eslint-enable no-var */
 }